requirements.txt
dnspython==2.7.0
pymongo==4.10.1
python-dotenv==1.0.1
